fixes #11955 - refactor automatic key acceptance
Instead of relying on orchestration, keys are accepted when the built URL
is called.  They are then removed when in a before_destroy callback.
This had the advantage of working with either image or normal
provisioning, and decreasing the complexity a bit. It's also probaby a
bit more secure, as the autosign file is never used.